shinycssloaders {shinycssloaders}	R Documentation

shinycssloaders.

Description

Add automatic loading animations to your Shiny outputs (e.g. plots, tables) by adding a single function call around the Shiny output function. Animations automatically show / hide when the output is recalculating / recalculated (or error is throws). You can choose from 8 different types of animations and you can also customize the color / size of the animation.

Details

For further reference on how to use the package, please refer to https://github.com/andrewsali/shinycssloaders

[Package shinycssloaders version 0.2.0 Index]