shinycssloaders-package {shinycssloaders}	R Documentation

shinycssloaders: Add Loading Animations to a 'shiny' Output While It's Recalculating

Description

When a 'Shiny' output (such as a plot, table, map, etc.) is recalculating, it remains visible but gets greyed out. Using 'shinycssloaders', you can add a loading animation ("spinner") to outputs instead. By wrapping a 'Shiny' output in 'withSpinner()', a spinner will automatically appear while the output is recalculating. You can also manually show and hide the spinner, or add a full-page spinner to cover the entire page. See the demo online at https://daattali.com/shiny/shinycssloaders-demo/.

Author(s)

Maintainer: Dean Attali daattali@gmail.com (ORCID) (Maintainer/developer of shinycssloaders since 2019)

Authors:

Andras Sali andras.sali@alphacruncher.hu (Original creator of shinycssloaders package)

Other contributors:

Luke Hass (Author of included CSS loader code) [contributor, copyright holder]

See Also

Useful links:

https://github.com/daattali/shinycssloaders

https://daattali.com/shiny/shinycssloaders-demo/

Report bugs at https://github.com/daattali/shinycssloaders/issues

[Package shinycssloaders version 1.1.0 Index]